# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Setu Health API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/v1/patients/search": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "required": true,
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mode": {
                      "type": "string",
                      "enum": [
                        "patientNo",
                        "phone",
                        "bn",
                        "en"
                      ]
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "facilityNo": {
                            "type": "string"
                          },
                          "nameBn": {
                            "type": "string"
                          },
                          "nameEn": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "sex": {
                            "type": "string",
                            "enum": [
                              "female",
                              "male",
                              "other"
                            ]
                          },
                          "birthDate": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "approxAgeYears": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "approxAgeMonths": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "approxAgeAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "phone": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "pattern": "^1[3-9]\\d{8}$"
                          },
                          "phoneOwner": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "guardian": {
                            "type": [
                              "object",
                              "null"
                            ],
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "relationship": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "name",
                              "relationship"
                            ]
                          },
                          "address": {
                            "type": "object",
                            "properties": {
                              "division": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "district": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "upazila": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "line": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              }
                            },
                            "required": [
                              "division",
                              "district",
                              "upazila",
                              "line"
                            ]
                          },
                          "hasNid": {
                            "type": "boolean"
                          },
                          "identityConfidence": {
                            "type": "string",
                            "enum": [
                              "verified",
                              "unverified",
                              "possible-duplicate",
                              "provisional"
                            ]
                          },
                          "linkedToId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "lastVisitAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "facilityNo",
                          "nameBn",
                          "nameEn",
                          "sex",
                          "birthDate",
                          "approxAgeYears",
                          "approxAgeMonths",
                          "approxAgeAt",
                          "phone",
                          "phoneOwner",
                          "guardian",
                          "address",
                          "hasNid",
                          "identityConfidence",
                          "linkedToId",
                          "lastVisitAt"
                        ]
                      }
                    },
                    "sharedPhone": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "phone": {
                          "type": "string",
                          "pattern": "^1[3-9]\\d{8}$"
                        },
                        "count": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "phone",
                        "count"
                      ]
                    }
                  },
                  "required": [
                    "mode",
                    "items",
                    "sharedPhone"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message_bn": {
                      "type": "string"
                    },
                    "message_en": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "canRequest": {
                      "type": "boolean"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "code"
                        ]
                      }
                    },
                    "existing": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message_bn",
                    "message_en"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/patients/match-preview": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nameBn": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "nameEn": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "sex": {
                    "type": "string",
                    "enum": [
                      "female",
                      "male",
                      "other"
                    ]
                  },
                  "dobMode": {
                    "type": "string",
                    "enum": [
                      "dob",
                      "age"
                    ]
                  },
                  "dob": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "ageYears": {
                    "type": "string",
                    "maxLength": 5
                  },
                  "ageMonths": {
                    "type": "string",
                    "maxLength": 4
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 30
                  },
                  "phoneOwner": {
                    "type": "string",
                    "enum": [
                      "self",
                      "guardian",
                      "family",
                      "other"
                    ]
                  },
                  "division": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "district": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "upazila": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "addressLine": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "guardian": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "maxLength": 120
                      },
                      "relationship": {
                        "type": "string",
                        "maxLength": 40
                      },
                      "idNo": {
                        "type": "string",
                        "maxLength": 30
                      }
                    }
                  },
                  "idType": {
                    "type": "string",
                    "enum": [
                      "none",
                      "nid",
                      "brn",
                      "passport"
                    ]
                  },
                  "idNo": {
                    "type": "string",
                    "maxLength": 30
                  }
                },
                "required": [
                  "nameBn",
                  "dobMode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "possible matches for an unsaved form",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candidates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "patient": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "facilityNo": {
                                "type": "string"
                              },
                              "nameBn": {
                                "type": "string"
                              },
                              "nameEn": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "sex": {
                                "type": "string",
                                "enum": [
                                  "female",
                                  "male",
                                  "other"
                                ]
                              },
                              "ageYears": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "ageApprox": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "id",
                              "facilityNo",
                              "nameBn",
                              "nameEn",
                              "sex",
                              "ageYears",
                              "ageApprox"
                            ]
                          },
                          "score": {
                            "type": "integer"
                          },
                          "strong": {
                            "type": "boolean"
                          },
                          "conflictCount": {
                            "type": "integer"
                          },
                          "isGuardian": {
                            "type": "boolean"
                          },
                          "canLink": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "patient",
                          "score",
                          "strong",
                          "conflictCount",
                          "isGuardian",
                          "canLink"
                        ]
                      }
                    }
                  },
                  "required": [
                    "candidates"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}